Fix bug when implicit operation was done, the result is wrong.
As it turns out, it is because I handled the implicit operation wrong by
making the operation the same as the explicit one. The difference
between explicit and implicit operation are, after implicit operation,
when the user enters another number, it became the second number. But,
with explicit operation, it should start a new calculation.

What I did wrong is I started a new calculation when I should be
updating the second number with implicit operation.